Rename `validate_string_concatenation.py` to `validate_unwanted_patterns.py`
