+++ scripts/validate_unwanted_patterns.py
